requirements.txt
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
colorclass==2.2.2
cx-Freeze==6.15.11
docopt==0.6.2
idna==3.6
macholib==1.16.2
numpy==1.26.2
packaging==23.2
Pillow==10.0.1
pip-upgrader==1.4.15
pipreqs==0.4.13
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
requests==2.31.0
six==1.16.0
terminaltables==3.1.10
urllib3==2.1.0
wxPython==4.2.1
yarg==0.1.9
